Re-home EET
